requirements.txt
# Michigan Lottery Results Analyzer - Requirements
# Python 3.10+ LLM-based Lottery Pattern Analysis with Smart Number Suggestions

# Core LLM and AI Frameworks
streamlit>=1.40.0
openai>=1.3.0
# Data Analysis and Visualization
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.6.0
plotly>=5.17.0
seaborn>=0.12.0

# Web Scraping and Data Collection
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium==4.20.0  # Fixed version for PDF scraper compatibility
webdriver-manager>=4.0.0
urllib3==2.2.1  # Fixed version to avoid zstandard compatibility issues

# Machine Learning (Optional)
scikit-learn>=1.2.0

# Utilities
python-dotenv>=1.1.0
networkx>=3.5.0
schedule>=1.2.0
python-dateutil>=2.9.0

# API and Web Development
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.28.0

# Data Storage
sqlite3-utils>=3.37.0
# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Statistical Analysis
scipy>=1.10.0
PyMuPDF>=1.24.0  # PDF text & image extraction (fitz)
pytesseract>=0.3.10  # OCR engine wrapper
pdf2image>=1.17.0  # PDF to image conversion for OCR
Pillow>=10.0.0  # Imaging support for OCR